Refactors describe command to reduce cyclomatic complexity
